merge/pull: --compact-summary
By default, "git merge" and "git pull" shows "git diff --stat @{1}"
to show the extent of the changes.  While it gives a good overview,
one thing that is missing in this output is which paths are created
and/or deleted.

Introduce "--compact-summary" option to these two commands that
tells it to instead show "git diff --compact-summary @{1}", which
gives the same diffstat but notes the created or deleted paths.

Signed-off-by: Junio C Hamano <gitster@pobox.com>